Fix exception when running within Azure web app
When running an Orchard Core instance within an Azure environment, when
a custom theme is enabled an exception prevents the front end from being
accessible. Below is a snipprt of the logged exception.

```
System.InvalidOperationException: Cannot find compilation library location for package 'Microsoft.NETCore.App'
```

To fix this issue (discussed here - https://github.com/OrchardCMS/OrchardCore/issues/1156)
a couple of properties need to be added to the csproj for the main site
project.

Fixes #3.